+++ Task4/src/task7.c
@@ -3,6 +3,6 @@ int main(){
     int *a = malloc(100);
     a[0] = 1;
     free(a);
-    printf ("value: %li\n", (int)a[0]);
+    printf ("value: %li\n", (long)a[0]);
     return 0;
 }
